+++ CHANGELOG.md
@@ -22,7 +22,7 @@ A small point release to fix a couple of important issues that slipped into 3.16
 
 My thanks to the following for helping with the Phaser 3 Examples, Docs and TypeScript definitions, either by reporting errors, fixing them or helping author the docs:
 
-@maretana @CipSoft-Components
+@maretana @CipSoft-Components @brian-lui
 
 ## Version 3.16.0 / 3.16.1 - Ishikawa - 5th February 2019